refactor: interfaces-api 로 패키지 재구성
