+++ src/main/java/kitchenpos/menu/interfaces/api/MenuDtoMapper.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.interfaces;
+package kitchenpos.menu.interfaces.api;
 
 import kitchenpos.menu.domain.MenuCommand;
