+++ src/main/java/kitchenpos/menu/interfaces/api/MenuRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.interfaces;
+package kitchenpos.menu.interfaces.api;
 
 import kitchenpos.menu.application.MenuFacade;
 import kitchenpos.menu.domain.MenuCommand;
